Use hooks for computer layout
